The ListPokemon page was created succesfully, with the components and the route
