+++ package.json
@@ -17,6 +17,7 @@
     "axios": "^1.7.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
     "sass": "^1.79.4"
   },
   "devDependencies": {
